Refactor into vision and video subpackages
