+++ src/ucl_open/video/__init__.py
@@ -1,0 +1,3 @@
+from ucl_open.video.video import ArducamOV9180, SpinnakerCamera, CameraModule, Camera
+
+__all__ = ["ArducamOV9180", "SpinnakerCamera", "CameraModule", "Camera"]
